add teachingBubble to PerceptorTab
